requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.1
python-multipart>=0.0.5
docker>=5.0.3
anthropic>=0.3.0
pydantic>=1.9.0
jwt>=1.3.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0